python-entra-mcp-server/pyproject.toml
[project]
name = "python-entra-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-identity>=1.25.2",
    "mcp[server]>=1.26.0",
    "msgraph-sdk>=1.55.0",
    "python-dotenv>=1.2.1",
    "starlette>=0.52.1",
    "uvicorn>=0.41.0",
]
